requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
django-unfold==0.66.0
django-extensions==3.2.3
python-decouple==3.8
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
django-filter==24.3
Pillow==10.4.0